app/pyproject.toml
[project]
name = "es-to-sqlite-migration"
version = "0.1.0"
description = "Migrate data from Elasticsearch to SQLite"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "elasticsearch>=8.0.0",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.11"]